!fixup rename to sdk.harness.JvmInitializer
