+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/DataflowBatchWorkerHarness.java
@@ -25,7 +25,7 @@ import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import org.apache.beam.runners.dataflow.options.DataflowWorkerHarnessOptions;
-import org.apache.beam.sdk.fn.BeamWorkerInitializers;
+import org.apache.beam.sdk.fn.JvmInitializers;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.util.BackOff;
 import org.apache.beam.sdk.util.BackOffUtils;
@@ -55,15 +55,15 @@ public class DataflowBatchWorkerHarness {
   /** Creates the worker harness and then runs it. */
   public static void main(String[] args) throws Exception {
     // Call user-defined initialization immediately upon starting, as is guaranteed in
-    // BeamWorkerInitializer
-    BeamWorkerInitializers.runOnStartup();
+    // JvmInitializer
+    JvmInitializers.runOnStartup();
     DataflowWorkerHarnessHelper.initializeLogging(DataflowBatchWorkerHarness.class);
     DataflowWorkerHarnessOptions pipelineOptions =
         DataflowWorkerHarnessHelper.initializeGlobalStateAndPipelineOptions(
             DataflowBatchWorkerHarness.class);
     DataflowBatchWorkerHarness batchHarness = new DataflowBatchWorkerHarness(pipelineOptions);
     DataflowWorkerHarnessHelper.configureLogging(pipelineOptions);
-    BeamWorkerInitializers.runBeforeProcessing(pipelineOptions);
+    JvmInitializers.runBeforeProcessing(pipelineOptions);
     batchHarness.run();
   }
