+++ sdks/java/core/src/main/java/org/apache/beam/sdk/harness/package-info.java
@@ -18,7 +18,7 @@
 
 /** Utilities for configuring worker environment. */
 @DefaultAnnotation(NonNull.class)
-package org.apache.beam.sdk.worker;
+package org.apache.beam.sdk.harness;
 
 import edu.umd.cs.findbugs.annotations.DefaultAnnotation;
 import edu.umd.cs.findbugs.annotations.NonNull;
